ZeroMQ Bugfixes, also added newest ZMQ source code to project
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20002 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: a8e6e3964560a00f4ec0a2ec3622feeb044fd22a

+++ server/wrapper/zeromq/wrapper.zeromq/src/zmq/Decoder.java
@@ -98,7 +98,7 @@ public class Decoder extends DecoderBase {
                 return false;
             }
             
-            int size = first;
+            int size = (int) first;
             if (size < 0) {
                 size = (0xFF) & first;
             }
@@ -106,7 +106,7 @@ public class Decoder extends DecoderBase {
             //  in_progress is initialised at this point so in theory we should
             //  close it before calling zmq_msg_init_size, however, it's a 0-byte
             //  message and thus we can treat it as uninitialised...
-            if (maxmsgsize >= 0 && size - 1 > maxmsgsize) {
+            if (maxmsgsize >= 0 && (long) (size - 1) > maxmsgsize) {
                 decoding_error ();
                 return false;
